% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_filtering.R
\name{dataset_filtering}
\alias{dataset_filtering}
\title{Data Filtering}
\usage{
dataset_filtering(
  data.path,
  guideline_file_name = "MEA_Spikes_ANA_R_guideline",
  sheet_used = 1,
  MinFR = 0,
  spike.sorting = FALSE,
  min.spike = 10,
  validation.index = "silhouette",
  cutoff.index = 0.5,
  nb.electrode = 64,
  save.rDATA = TRUE,
  filename = "MEA_DATA2ANALYZE"
)
}
\arguments{
\item{data.path}{The path directory to the folder containing the data and the guideline.}

\item{guideline_file_name}{The name of the excel file containing the guideline (by default "MEA_Spikes_ANA_R_guideline")}

\item{sheet_used}{Number indicating the excel sheet where the guideline is located (by default 1)}

\item{MinFR}{Minimum firing rate to consider (by default 0)}

\item{spike.sorting}{Boolean, if TRUE a spike sorting is performed (by default FALSE)}

\item{min.spike}{Minimum number of spikes in the considered time window to be used during spike sorting (by default 10)}

\item{validation.index}{The method used to determine the optimal number of clusters (by default "silhouette")}

\item{cutoff.index}{The minimum threshold for the selection method in
\code{validation.index} to determine the optimal spike sorting cluster number (by default 0.5)}

\item{nb.electrode}{Number which specifies the number of electrodes located on the MEA chip (by default 64)}

\item{save.rDATA}{Boolean, if TRUE (by default) the filtered data are saved in the directory specified in \code{data.path}}

\item{filename}{Character, the name of the .Rdat file in which filtered data will
be saved if \code{save.rDATA = TRUE} (by default "MEA_DATA2ANALYZE")}
}
\value{
An object of class 'MEASpikeR' which will be the input of all other functions of the MEASpikeR package.
if \code{save.rDATA = TRUE}, a ".Rdat" file called by default "MEA_DATA2ANALYZE" is saved
in the \code{data.path} directory
}
\description{
Filters the data according to the conditions of the guideline
}
\examples{
\dontrun{
dataset_filtering(data.path = "C:/Users/prenom.nom/Documents/data",
guideline_file_name = "MEA_Spikes_ANA_R_guideline")
}

}
\references{
Dalle C., Martet G., Lejeune FX., Rebola N., Kaddouri Y., Whitmarsh S.(2023). \strong{\emph{MEASpikeR}: a new package
for spike analysis and visualization of \emph{in vitro} MEA data}.
}
\author{
François-Xavier Lejeune, Gaspard Martet, Carine Dalle
}
