% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firing_rate.R
\name{firing_rate}
\alias{firing_rate}
\title{Firing Rate and Mean Firing Rate}
\usage{
firing_rate(object = NA, filt.data.path = NA, working_directory = NA)
}
\arguments{
\item{object}{Object of class MEASpikeR generated by the \code{dataset_filtering} function}

\item{filt.data.path}{Path to "MEA_DATA2ANALYZE.Rdat" file generated by the \code{dataset_filtering} function}

\item{working_directory}{Character specifying the working directory where we want to save the output excel file}
}
\value{
An excel file with the firing rate and the mean firing rate for each condition located in the working directory
specified in the \code{working_directory} argument
}
\description{
Firing Rate and Mean Firing Rate
}
\examples{
\dontrun{
out.MEA <- MEASpikeR::dataset_filtering(data.path = "C:/Users/prenom.nom/Documents/data",
save.rDATA = TRUE)

# if you want to work with the MEASpikeR object
MEASpikeR::firing_rate(
object = out.MEA,
working_directory = "C:/Users/prenom.nom/Documents"
)

# if you want to work with the saved Rdat file
MEASpikeR::firing_rate(
filt.data.path = "C:/Users/prenom.nom/Documents/data/MEA_DATA2ANALYZE.Rdat",
working_directory = "C:/Users/prenom.nom/Documents"
)}
}
\references{
Dalle C., Martet G., Lejeune FX., Rebola N., Kaddouri Y., Whitmarsh S.(2023). \strong{\emph{MEASpikeR}: a new package
for spike analysis and visualization of \emph{in vitro} MEA data}.
}
\author{
François-Xavier Lejeune, Gaspard Martet, Carine Dalle, Stephen Whitmarsh
}
