% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISI.R
\name{ISI}
\alias{ISI}
\title{ISI Histogram}
\usage{
ISI(
  object = NA,
  filt.data.path = NA,
  output_directory = NA,
  col_hist = "darkgreen",
  bin.width = 50,
  interval.max = 300,
  format.save = "both"
)
}
\arguments{
\item{object}{Object of class MEASpikeR generated by the \code{dataset_filtering} function}

\item{filt.data.path}{Path to the ".Rdat" file generated
by the \code{dataset_filtering} function}

\item{output_directory}{Character specifying the directory where we want to save the output folder}

\item{col_hist}{Character specifying the color of the histogram (by default "darkgreen")}

\item{bin.width}{Number, parameter of the histogram (by default 50)}

\item{interval.max}{Number, parameter of the histogram (by default 300)}

\item{format.save}{Character specifying the format in which you want to save the graphics ("both",
"one_page" or "several_pages"). if the value is "one_page",  it will display all the graphics for a condition on the
same page in a table with 8 rows and 8 columns and save them in '.png' format.
If the value is "several_pages",
it will diplay every graphics in a ".pdf" document with 12 graphics per page.
if the value is "both", it will perform both methods.}
}
\value{
A PDF file and/or an image with the ISI histogram for
each channel is generated for each condition. All files are located in a folder
named ‘ISIh_outfiles’ with the date and time as a prefix. In the PDF files,
there are 12 graphics per page, and images represent an 8x8 matrix of graphics.
Excluded channels are represented by a cross, and channels with no data are represented by a blank.
}
\description{
Generate ISI histogram pdf files
}
\examples{
\dontrun{
out.MEA <- MEASpikeR::dataset_filtering(data.path = "C:/Users/prenom.nom/Documents/data",
guideline_name_file = "MEA_Spikes_ANA_R_guideline",
save.rDATA = TRUE)

# if you want to work with the MEASpikeR object
MEASpikeR::ISI(
object = out.MEA,
output_directory = "C:/Users/prenom.nom/Documents"
)

# if you want to work with the saved Rdat file
MEASpikeR::ISI(
filt.data.path = "C:/Users/prenom.nom/Documents/data/MEA_DATA2ANALYZE.Rdat",
output_directory = "C:/Users/prenom.nom/Documents"
)
}
}
\references{
Dalle C., Martet G., Kaddouri Y., Rebola N., Whitmarsh S., Lejeune FX.(2023). \strong{\emph{MEASpikeR}: a new package
for spike analysis and visualization of \emph{in vitro} MEA data}.
}
\author{
François-Xavier Lejeune, Gaspard Martet, Carine Dalle
}
