% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spike_waveform.R
\name{spike_waveform}
\alias{spike_waveform}
\title{Spike Waveform}
\usage{
spike_waveform(
  object = NA,
  filt.data.path = NA,
  output_directory = NA,
  spike_col = "forestgreen",
  scale = "free",
  mean_trace = TRUE,
  col_mean_trace = "black",
  color_viridis = FALSE,
  format.save = "both"
)
}
\arguments{
\item{object}{Object of class MEASpikeR generated by the \code{dataset_filtering} function}

\item{filt.data.path}{Path to "MEA_DATA2ANALYZE.Rdat" file generated
by the \code{dataset_filtering} function}

\item{output_directory}{Character specifying the directory where you want to save the output folder}

\item{spike_col}{Character specifying the color parameter for spikes
in case spike sorting is FALSE (cluster_id = "none"), (by default "forestgreen")}

\item{scale}{Character specifying the scale (µV) for the y-axis ("free"/"fixed"),
(by default "free")}

\item{mean_trace}{Boolean, if TRUE the average curve is plotted (by default TRUE)}

\item{col_mean_trace}{Character specifying the color of the average curve (by default "black")}

\item{color_viridis}{Boolean, if TRUE the viridis color palette is used
for clusters (by default FALSE)}

\item{format.save}{Character specifying the format in which you want to save the graphics ("both",
"one_page" or "several_pages"). if the value is "one_page",  it will display all the graphics for a condition on the
same page in a table with 8 rows and 8 columns and save them in '.png' format.
If the value is "several_pages",
it will diplay every graphics in a ".pdf" document with 12 graphics per page.
if the value is "both", it will perform both methods.}
}
\value{
A PDF file and/or an image with the overlapped spike display/plot for
each channel is generated for each condition. All files are located in a folder
named ‘outfile_waveform’ with the date and time as a prefix. In the PDF files,
there are 12 graphics per page, and images represent an 8x8 matrix of graphics.
Excluded channels are represented by a cross, and channels with no data are represented by a blank.
}
\description{
Display of overlapped events for each electrode. In the case of spike sorting,
display of colored events according to their assignment to a cluster
}
\examples{
\dontrun{
out.MEA <- MEASpikeR::dataset_filtering(data.path = "C:/Users/prenom.nom/Documents/data",
guideline_name_file = "MEA_Spikes_ANA_R_guideline",
save.rDATA = TRUE)

# if you want to work with the MEASpikeR object
MEASpikeR::spike_waveform(
object = out.MEA,
output_directory = "C:/Users/prenom.nom/Documents"
)

# if you want to work with the saved Rdat file
MEASpikeR::spike_waveform(
filt.data.path = "C:/Users/prenom.nom/Documents/data/MEA_DATA2ANALYZE.Rdat",
output_directory = "C:/Users/prenom.nom/Documents"
)}
}
\references{
Dalle C., Martet G., Kaddouri Y., Rebola N., Whitmarsh S., Lejeune FX.(2023). \strong{\emph{MEASpikeR}: a new package
for spike analysis and visualization of \emph{in vitro} MEA data}.
}
\author{
François-Xavier Lejeune, Gaspard Martet, Carine Dalle
}
