% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikeraster_heatmap.R
\name{spikeraster_heatmap}
\alias{spikeraster_heatmap}
\title{Spike Rasters and Heatmaps}
\usage{
spikeraster_heatmap(
  object = NA,
  filt.data.path = NA,
  working_directory = NA,
  scale = "free",
  end_record_color = "red",
  spike_col = "black",
  add.burst = TRUE,
  burst.detection.method = "MI",
  max_ISI_start_burst = 170,
  max_ISI_end_burst = 300,
  min_IBI = 50,
  min_spike_in_burst = 5,
  min_burst_duration = 50,
  mypal = "viridis2",
  col.cell.border = "gray85"
)
}
\arguments{
\item{object}{Object of class MEASPIRs generated by the \code{dataset_filtering} function}

\item{filt.data.path}{Path to the "MEA_DATA2ANALYZE.Rdat" file generated
by the \code{dataset_filtering} function.}

\item{working_directory}{Character specifying the working directory where we want to save the output pdf file}

\item{scale}{Character specifying the scale ("free"/"fixed")(by default "free")}

\item{end_record_color}{Character specifying the color of vertical bar at the end of recording (by default "red") for the raster plots}

\item{spike_col}{Character specifying the color of the spikes of the raster plots (by default "black")}

\item{add.burst}{Boolean, if TRUE, it displays bursts in color in the raster plot}

\item{burst.detection.method}{Character specifying the choice of
the burst detection method ("MI" or "Chen")(by default "MI")}

\item{max_ISI_start_burst}{Parameter settings for burst detection (for MI) (by default 170)}

\item{max_ISI_end_burst}{Parameter settings for burst detection (for MI) (by default 300)}

\item{min_IBI}{Parameter settings for burst detection (for MI) (by default 50)}

\item{min_spike_in_burst}{Parameter settings for burst detection (for MI) (by default 5)}

\item{min_burst_duration}{Parameter settings for burst detection (for MI) (by default 50)}

\item{mypal}{Character specifying the color palette for the heatmap (by default "viridis2")}

\item{col.cell.border}{Character specifying the color delimiting the cells in the heatmap (by default "gray85")}
}
\value{
A PDF file displaying raster plots and heatmaps located in the working directory
}
\description{
Create raster plots and heatmaps
}
\examples{
\dontrun{
out.MEA <- MEASPIRs::dataset_filtering(data.path = "C:/Users/prenom.nom/Documents/data",
save.rDATA = TRUE)

# if you want to work with the MEASPIRs object
MEASPIRs::spikeraster_heatmap(
object = out.MEA,
working_directory = "C:/Users/prenom.nom/Documents"
)

# if you want to work with the saved Rdat file
MEASPIRs::spikeraster_heatmap(
filt.data.path = "C:/Users/prenom.nom/Documents/data/MEA_DATA2ANALYZE.Rdat",
working_directory = "C:/Users/prenom.nom/Documents"
)}
}
\references{
Dalle C., Martet G., Lejeune FX., Rebola N., Kaddouri Y., Whitmarsh S.(2023). \strong{\emph{MEASPIRs}: a new package
for spike analysis and visualization of \emph{in vitro} MEA data}.
}
\author{
François-Xavier Lejeune, Gaspard Martet, Carine Dalle, Stephen Whitmarsh
}
