% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spike_rate_plot.R
\name{spike_rate_plot}
\alias{spike_rate_plot}
\title{Spike Rate Plot}
\usage{
spike_rate_plot(
  object = NA,
  filt.data.path = NA,
  output_directory = NA,
  spike_col = "forestgreen",
  plot = "line"
)
}
\arguments{
\item{object}{Object of class MEASpikeR generated by the \code{dataset_filtering} function}

\item{filt.data.path}{Path to the ".Rdat" file generated by the \code{dataset_filtering} function}

\item{output_directory}{Character specifying the directory where we want to save the output excel file}

\item{spike_col}{Character specifying the color for the spikes with no spike sorting (by default "forestgreen")}

\item{plot}{Character specifying the type of plot ("needle" or "line") (by default "line")}
}
\value{
An excel file called "AWSDR" with date and time as suffix is created with as many sheets as there are conditions,
where each sheet contains a table with two columns.
The first column indicates time, and the second column indicates the firing rate.
Additionally, there is a graphic per condition representing the firing rate (Hz) as a function of time (s).
All these files are saved in the output directory specified in the 'output_directory' argument.
}
\description{
Spike Rate Plot
}
\examples{
\dontrun{
out.MEA <- MEASpikeR::dataset_filtering(data.path = "C:/Users/prenom.nom/Documents/data",
guideline_name_file = "MEA_Spikes_ANA_R_guideline",
save.rDATA = TRUE)

# if you want to work with the MEASpikeR object
MEASpikeR::spike_rate_plot(
object = out.MEA,
output_directory = "C:/Users/prenom.nom/Documents"
)

# if you want to work with the saved Rdat file
MEASpikeR::spike_rate_plot(
filt.data.path = "C:/Users/prenom.nom/Documents/data/MEA_DATAFILTERED.Rdat",
output_directory = "C:/Users/prenom.nom/Documents"
)
}
}
\references{
Dalle C., Martet G., Kaddouri Y., Rebola N., Whitmarsh S., Lejeune FX.(2023). \strong{\emph{MEASpikeR}: a new package
for spike analysis and visualization of \emph{in vitro} MEA data}.
}
\author{
François-Xavier Lejeune, Gaspard Martet, Carine Dalle
}
